Rename "store" command to "vts"
It has become clear that what we previously referred to as the
"endorsement store node" would have a wider set of responsibilities than
just handling endorsements. As such, that name was misleading and a new
name -- Veraison Trusted Services, or VTS (pronounced "vitis") -- was
chosen as more representative of the node's purpose.
