--- cmd/store/Makefile
@@ -1,12 +1,0 @@
-# Copyright 2021 Contributors to the Veraison project.
-# SPDX-License-Identifier: Apache-2.0
-
-GOPKG := github.com/veraison/cmd/store
-SRCS := main.go
-CMD := store
-
-include ../../mk/cmd.mk
-include ../../mk/common.mk
-include ../../mk/lint.mk
-include ../../mk/pkg.mk
-include ../../mk/test.mk
