+++ cmd/vts/Makefile
@@ -1,0 +1,12 @@
+# Copyright 2021 Contributors to the Veraison project.
+# SPDX-License-Identifier: Apache-2.0
+
+GOPKG := github.com/veraison/cmd/trustedservices
+SRCS := main.go
+CMD := vts
+
+include ../../mk/cmd.mk
+include ../../mk/common.mk
+include ../../mk/lint.mk
+include ../../mk/pkg.mk
+include ../../mk/test.mk
