+++ cmd/vts/README.md
@@ -1,0 +1,4 @@
+# Veraison Trusted Services
+
+This is an implementation of the Veraison Trusted Services node taht provides
+the endorsements store and attestation scheme management.
